moved helper tasks out of trajectory control
