+++ firmware/src/goto.rs
@@ -1,8 +1,8 @@
 use crate::math::SO2;
 use crate::read_robot_config_from_sd::RobotConfig;
-use crate::sdlog::TrajControlLog;
-use crate::trajectory_control::{STOP_ALL, with_sdlogger};
-use crate::trajectory_control::{
+use crate::sdlog::{TrajControlLog, with_sdlogger};
+use crate::orchestrator_signal::STOP_ALL;
+use crate::control_types::{
     DiffdriveCascade, DiffdriveControllerCascade, PointCascade, TrajectoryResult,
 };
 use crate::robotstate;
