+++ firmware/src/orchestrator_signal.rs
@@ -44,6 +44,10 @@ pub static STOP_POSE_EST_SIG: Signal<Raw, ()> = Signal::new();
 pub static TRAJ_PAUSE_SIG: Signal<Raw, bool> = Signal::new();
 pub static TRAJ_RESUME_SIG: Signal<Raw, bool> = Signal::new();
 
+/// Legacy global stop flag, checked in execute loops and goto.rs.
+/// Moved from trajectory_control.rs to consolidate lifecycle control.
+pub static STOP_ALL: portable_atomic::AtomicBool = portable_atomic::AtomicBool::new(false);
+
 pub fn decode_functionality_select_command(payload: &[u8], robot_id: u8) -> Option<u8> {
     // Check frame format: should be 4 bytes with specific header
     if payload.len() != (LEN_FUNC_SELECT_CMD as usize) || payload[0] != 0x3C {
